requirements.txt
# requirements.txt

discord.py>=2.3.0
python-dotenv>=1.0.0
pymongo>=4.6.0
aiohttp>=3.9.0
pytz>=2024.1
matplotlib>=3.8.0
numpy>=1.26.0
seaborn>=0.13.0
pandas>=2.2.0
APScheduler>=3.10.0
